requirements.txt
asgiref==3.8.1
attrs==23.2.0
cachetools==5.3.3
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
configloader==1.0.1
cryptography==42.0.8
defusedxml==0.8.0rc2
Deprecated==1.2.14
Django==5.0.6
django-allauth==0.63.3
django-anymail==11.0
django-mailgun-mime==0.1.7
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
et-xmlfile==1.1.0
GoDaddyPy==2.5.1
google-auth==2.31.0
httpie==3.2.3
idna==3.7
markdown-it-py==3.0.0
mdurl==0.1.2
msal==1.29.0
multidict==6.0.5
oauthlib==3.2.2
openpyxl==3.1.5
pillow==10.3.0
psycopg2-binary==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
Pygments==2.18.0
PyJWT==2.8.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python3-openid==3.2.0
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==13.7.1
rsa==4.9
rustici-software-cloud-v2==4.0.0
setuptools==70.3.0
six==1.16.0
social-auth-app-django==5.4.1
social-auth-core==4.5.4
sqlparse==0.5.0
typing_extensions==4.12.1
urllib3==2.2.2
wrapt==1.16.0
